# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PERSONAL-3 Alpaca Trading System — Operator API
  version: 1.1.0
  license:
    name: Proprietary — PERSONAL-3 operator only
    identifier: LicenseRef-PERSONAL-3-Proprietary
  description: |
    Ганц operator-т зориулсан арилжааны системийн REST API.

    **Бүх хариунд хүчинтэй гурван үл хөдлөх дүрэм:**
    1. Мөнгөн утга бүр **тэмдэгт мөр** хэлбэрийн fixed-point decimal
       (`"10234.56"`). Float ХЭЗЭЭ Ч биш (AC-25).
    2. Timestamp бүр UTC, RFC 3339, `Z` төгсгөлтэй (AC-26).
    3. Өгөгдөлтэй хариу бүр `source` талбартай ба нэг хариу дотор
       ХОЛИМОГ `source` байхыг хориглоно (AC-20, AC-21).

    **Аюулгүйн гэрээ:** энэ API-д Alpaca руу order илгээх ЦОРЫН ГАНЦ зам нь
    `POST /orders/manual` ба approval queue-ийн `approve` хоёр. Хоёул дотооддоо
    Risk Agent → Execution модулиар дамжина. LLM нь энэ API-д огт хүрэхгүй —
    LLM зөвхөн Tool Contract (энэ файлын 3-р баримт) руу ханддаг.
servers:
  - url: http://localhost:8000/api/v1
    description: DEV (paper)
  - url: https://{host}/api/v1
    description: UAT / Prod
    variables:
      host:
        default: trader.internal
...
security:
  - operatorSession: []
paths:
...
  /approvals:
    get:
      operationId: getApprovals
      tags: [approvals]
      summary: Зөвшөөрөл хүлээж буй саналууд
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, expired, all]
            default: pending
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalsEnvelope"
  /approvals/{approvalId}/approve:
    post:
      operationId: postApprovalsByApprovalIdApprove
      tags: [approvals]
      summary: Хүлээгдэж буй саналыг зөвшөөрөх
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version]
              properties:
                expected_version:
                  type: integer
                  description: |
                    Optimistic lock. `GET /approvals`-аас авсан `version`.
                    Зөрвөл 409 — TTL дуусаад `expired` болсон мөрийг
                    зөвшөөрөх боломжгүй (AC-6).
                note:
                  type: string
                  maxLength: 1000
      responses:
        "202":
          description: Зөвшөөрөгдөв, Execution-д дамжив
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderAccepted"
        "409":
          description: Мөр аль хэдийн шийдэгдсэн / `expired` / версийн зөрүү
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /approvals/{approvalId}/reject:
    post:
      operationId: postApprovalsByApprovalIdReject
      tags: [approvals]
      summary: Хүлээгдэж буй саналыг татгалзах (буцаах боломжгүй)
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version, reason]
              properties:
                expected_version:
                  type: integer
                reason:
                  type: string
                  minLength: 1
                  maxLength: 1000
      responses:
        "200":
          description: Татгалзав. Энэ санал ХЭЗЭЭ Ч илгээгдэхгүй.
        "409":
          description: Мөр аль хэдийн шийдэгдсэн эсвэл версийн зөрүү
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    operatorSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Ганц operator-ийн session cookie (HttpOnly, Secure, SameSite=Strict).
        Олон хэрэглэгчийн role model v1-д БАЙХГҮЙ (спек A-6).
...
  schemas:
    Money:
      type: string
      pattern: '^-?[0-9]+\.[0-9]{2}$'
      description: Fixed-point decimal тэмдэгт мөр. Float ХЭЗЭЭ Ч биш (AC-25).
      examples: ["10234.56", "-42.00"]
    Quantity:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,9})?$'
      description: Ширхэг. Crypto-д бутархай байж болох тул decimal тэмдэгт мөр.
    Utc:
      type: string
      format: date-time
      pattern: 'Z$'
      description: RFC 3339, ЗААВАЛ UTC (`Z`). Локал бүс хориотой (AC-26).
    Source:
      type: string
      enum: [alpaca_live, alpaca_paper, backtest]
      description: |
        Энэ хариун дахь БҮХ өгөгдлийн гарал. Нэг хариу дотор холимог утга
        БАЙХГҮЙ — холих шаардлагатай бол тусдаа хүсэлт (AC-20, AC-21).
    SystemState:
      type: string
      enum: [active, winding_down, halted]
    Origin:
      type: string
      enum: [research_agent, auto_tuning, manual_operator, external]
      description: |
        Order / позицийн гарал үүсэл (AC-29). `external` = локал `orders`-д
        таарах бичлэг ОЛДООГҮЙ (Alpaca UI-аас нээсэн г.м.). ХЭЗЭЭ Ч
        таамаглахгүй — таарахгүй бол `external`.
    RiskDecision:
      type: string
      enum: [APPROVE, REJECT, ESCALATE_TO_HUMAN]
    OrderSide:
      type: string
      enum: [buy, sell]
    OrderType:
      type: string
      enum: [market, limit, stop, stop_limit, bracket]
    TimeInForce:
      type: string
      enum: [day, gtc, ioc, fok]
    Envelope:
      type: object
      required: [source, as_of, stale, system_state]
      properties:
        source:
          $ref: "#/components/schemas/Source"
        as_of:
          $ref: "#/components/schemas/Utc"
        stale:
          type: boolean
          default: false
          description: |
            Өгөгдөл `STALE_AFTER_SECONDS`-оос хуучин. Талбарыг ОРХИХГҮЙ —
            ил `false` бичнэ (хавсралт 10 §2.4: «байхгүй» нь таамаглах цонх).
        system_state:
          $ref: "#/components/schemas/SystemState"
...
    Order:
      type: object
      required: [id, client_order_id, symbol, side, qty, order_type, status, origin, submitted_at]
      properties:
        id:
          type: string
          format: uuid
          description: Локал ID
        broker_order_id:
          type: [string, "null"]
          description: Alpaca-ийн ID; илгээгдэх хүртэл `null`
        client_order_id:
          type: string
          description: Idempotency түлхүүрээс гаргасан детерминистик ID
        symbol:
          type: string
        side:
          $ref: "#/components/schemas/OrderSide"
        qty:
          $ref: "#/components/schemas/Quantity"
        filled_qty:
          $ref: "#/components/schemas/Quantity"
        order_type:
          $ref: "#/components/schemas/OrderType"
        limit_price:
          $ref: "#/components/schemas/Money"
        stop_price:
          $ref: "#/components/schemas/Money"
        time_in_force:
          $ref: "#/components/schemas/TimeInForce"
        status:
          type: string
          enum: [pending_risk, pending_approval, rejected, accepted, partially_filled, filled, canceled, expired, failed]
        origin:
          $ref: "#/components/schemas/Origin"
        origin_detail:
          type: [string, "null"]
        decision_id:
          type: [string, "null"]
          format: uuid
          description: Холбогдох `agent_decisions` мөр
        submitted_at:
          $ref: "#/components/schemas/Utc"
        filled_at:
          oneOf:
            - $ref: "#/components/schemas/Utc"
            - type: "null"
          description: Биелэх хүртэл `null`
...
    OrderAccepted:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          order:
            id: 9a8b7c6d-5e4f-4a3b-8c2d-1e0f9a8b7c6d
            broker_order_id: null
            client_order_id: p3-9a8b7c6d
            symbol: MSFT
            side: sell
            qty: "12"
            filled_qty: "0"
            order_type: market
            time_in_force: day
            status: accepted
            origin: manual_operator
            origin_detail: operator
            decision_id: null
            submitted_at: "2026-09-16T14:29:51Z"
          risk:
            decision: APPROVE
            reason: null
            evaluated_at: "2026-09-16T14:29:51Z"
            checks:
              - rule: system_state
                passed: true
                limit_name: SYSTEM_STATE
                limit_value: active
                actual_value: active
              - rule: order_notional
                passed: true
                limit_name: MAX_ORDER_NOTIONAL
                limit_value: "5000.00"
                actual_value: "4870.80"
              - rule: wind_down_direction
                passed: true
                detail: "Позиц хаах чиглэл — winding_down үед ч зөвшөөрөгдөнө"
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [order, risk]
          properties:
            order:
              $ref: "#/components/schemas/Order"
            risk:
              $ref: "#/components/schemas/RiskEvaluation"
    RiskEvaluation:
      type: object
      required: [decision, checks, evaluated_at]
      properties:
        decision:
          $ref: "#/components/schemas/RiskDecision"
        reason:
          type: [string, "null"]
        evaluated_at:
          $ref: "#/components/schemas/Utc"
        checks:
          type: array
          description: >-
            Шалгалт бүрийн үр дүн — юу унасныг, ямар хязгаартай харьцуулснаар
            нь ил харуулна.
          items:
            type: object
            required: [rule, passed]
            properties:
              rule:
                type: string
                enum:
                  - system_state
                  - restricted_symbol
                  - position_pct
                  - total_exposure_pct
                  - daily_loss_limit
                  - pdt_day_trades
                  - order_notional
                  - price_sanity
                  - wind_down_direction
              passed:
                type: boolean
              limit_name:
                type: string
                description: "config-ийн нэр, жишээ: `MAX_POSITION_PCT`"
              limit_value:
                type: string
              actual_value:
                type: string
              detail:
                type: string
    Approval:
      type: object
      required: [id, version, state, decision_id, proposed_order, risk, created_at, expires_at]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        state:
          type: string
          enum: [pending, approved, rejected, expired]
        decision_id:
          type: string
          format: uuid
        proposed_order:
          $ref: "#/components/schemas/ProposedOrder"
        risk:
          $ref: "#/components/schemas/RiskEvaluation"
        created_at:
          $ref: "#/components/schemas/Utc"
        expires_at:
          allOf:
            - $ref: "#/components/schemas/Utc"
          description: "`created_at + APPROVAL_TTL`. Дараа нь автоматаар `expired` (AC-6)."
        resolved_at:
          oneOf:
            - $ref: "#/components/schemas/Utc"
            - type: "null"
          description: Шийдэгдэх хүртэл `null`
        resolution_reason:
          type: [string, "null"]
    ProposedOrder:
      type: object
      required: [symbol, side, qty, order_type, rationale, grounded_in, provider, model]
      properties:
        symbol:
          type: string
        side:
          $ref: "#/components/schemas/OrderSide"
        qty:
          $ref: "#/components/schemas/Quantity"
        order_type:
          $ref: "#/components/schemas/OrderType"
        limit_price:
          $ref: "#/components/schemas/Money"
        stop_price:
          $ref: "#/components/schemas/Money"
        estimated_notional:
          $ref: "#/components/schemas/Money"
        rationale:
          type: string
        grounded_in:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
            description: "`tool_calls.id`"
        provider:
          type: string
        model:
          type: string
    ApprovalsEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          approvals:
            - id: 2d3e4f50-6172-4839-a0b1-c2d3e4f50617
              version: 1
              state: pending
              decision_id: 3c9d8e7f-6a5b-4c3d-8e1f-0a9b8c7d6e5f
              proposed_order:
                symbol: NVDA
                side: buy
                qty: "25"
                order_type: limit
                limit_price: "402.00"
                estimated_notional: "10050.00"
                rationale: >-
                  20/50 EMA огтлолцол + өмнөх 3 өдрийн дундажаас 1.8 дахин их
                  эзлэхүүн. Эх өгөгдөл нь доорх tool call-уудад бүрэн байна.
                grounded_in:
                  - 11111111-2222-4333-8444-555555555555
                  - 66666666-7777-4888-8999-aaaaaaaaaaaa
                provider: claude-mcp
                model: claude-opus-5
              risk:
                decision: ESCALATE_TO_HUMAN
                reason: "Notional 10050.00 > MAX_ORDER_NOTIONAL 5000.00"
                evaluated_at: "2026-09-16T14:22:10Z"
                checks:
                  - rule: order_notional
                    passed: false
                    limit_name: MAX_ORDER_NOTIONAL
                    limit_value: "5000.00"
                    actual_value: "10050.00"
                  - rule: position_pct
                    passed: true
                    limit_name: MAX_POSITION_PCT
                    limit_value: "10"
                    actual_value: "9.6"
              created_at: "2026-09-16T14:22:10Z"
              expires_at: "2026-09-16T14:37:10Z"
              resolved_at: null
              resolution_reason: null
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [approvals]
          properties:
            approvals:
              type: array
              items:
                $ref: "#/components/schemas/Approval"
...
    Problem:
      description: RFC 9457 Problem Details.
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - confirmation_required
            - system_halted
            - winding_down_increase_blocked
            - risk_rejected
            - grounding_failed
            - idempotency_conflict
            - breaker_still_tripped
            - already_active
            - approval_expired
            - version_conflict
            - broker_unavailable
            - broker_rejected
            - provider_unavailable
            - not_found
            - invalid_request
        detail:
          type: string
        broker_code:
          type: string
          description: >-
            Зөвхөн `broker_rejected` үед. Alpaca-ийн алдааны код
            (ж: `42210000` — wash trade). Broker ХАРИУЛСАН гэдгийн нотолгоо.
        errors:
          type: array
          description: Зөвхөн `invalid_request` үед. Унасан талбар тус бүр.
          items:
            type: object
            required: [loc, msg]
            properties:
              loc:
                type: string
              msg:
                type: string
        risk:
          $ref: "#/components/schemas/RiskEvaluation"
        confirmation:
          type: object
          description: >-
            Зөвхөн `confirmation_required` үед. Хоёр шаттай баталгаажуулалтын
            сорилт.
          required: [token, prompt, expires_at]
          properties:
            token:
              type: string
            prompt:
              type: string
              description: |
                Operator-т харуулах тодорхой текст. Үйлдэл тус бүрд ӨӨР —
                хоёр товчийг андуурахаас сэргийлнэ (AC-38).
            expires_at:
              $ref: "#/components/schemas/Utc"
        breaker_metrics:
          type: array
          items:
            type: object
            additionalProperties: true